Derive frame times from the exposure indexes, not the per-channel file
Found while opening AB_0032/2024-07-24/1: that session's per-channel
svdTemporalComponents.timestamps.npy does not match its own V.

  channel  V frames   timestamps.npy   frameTimes[indexes==1]
  blue       194264   194678 (+414)    194264  correct
  violet     194631   194677  (+46)    194631  correct
  corr       194263   194263           n/a

The frame bookkeeping (frameTimes.timestamps.npy + <channel>Frames.indexes.npy)
is self-consistent and gives exactly the right count, which is also how
Pipelines/widefield/hemoCorrect.m derives its time base:
tb = frameTimes(bfIdx == 1).

load_uvt now prefers that route and falls back to the per-channel file, so the
session opens with correctly aligned times instead of tripping the movie
viewer's length check. A single trailing extra sample is still dropped
silently; a larger mismatch is truncated and warned about loudly, since it
means the times may be misaligned rather than merely one short. AB_0004 is
unaffected (all three channels already agreed).

Worth knowing: trusting the per-channel file on this session would have
misaligned the movie against behavior by up to 414 frames, about 12 s, with
nothing obviously wrong on screen.

Also:
- bandpass_filt was never exported from the package top level. Fixed.
- The movie viewer no longer promotes its spare copy of V to float64;
  bandpass_filt promotes internally anyway, and this saves 155 MB on a
  92-minute session.
- Launcher: --only picks a subset of viewers (and skips the work the others
  would need), tiling adapts to 1/2/4 windows, opto sessions get a laser
  power staircase trace from laserOnTimes/laserOffTimes/laserPowers, and the
  movie's reference trace now comes from the arrays actually displayed rather
  than from a different channel.

424 tests passing.

+++ src/widefield/gui/movie_with_traces.py
@@ -174,7 +174,10 @@ def _build():
             self._flat_u = flatten_u(self._u[..., :nsv]).astype(np.float32, copy=False)
             # Keep the unfiltered components: the band-pass control re-derives _v32 from these,
             # so filtering is always a single pass from the original rather than compounding.
-            self._v_raw = np.asarray(self._v[:nsv], dtype=np.float64)
+            # Left in the input dtype (float32 off disk) rather than promoted: bandpass_filt
+            # promotes internally anyway, and forcing float64 here doubled the resident size to
+            # 311 MB on a 92-minute session for no change in the result.
+            self._v_raw = np.asarray(self._v[:nsv])
             self._v32 = np.asfortranarray(self._v[:nsv], dtype=np.float32)
             self._block_start = -1
             self._block: np.ndarray | None = None
